% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fetch_study}
\alias{fetch_study}
\title{Fetch info for a clinical study}
\usage{
fetch_study(eudract, cache_file = NULL)
}
\arguments{
\item{eudract}{an eudract id}

\item{cache_file}{RDS file used to cache data}
}
\value{
a list of results
}
\description{
Fetch info for a clinical study
}
\examples{
fetch_study("2015-001314-10")
}
