% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{info}
\alias{info}
\title{Get info for 1 study}
\usage{
info(eudract)
}
\arguments{
\item{eudract}{an eudract id}
}
\value{
a list of results
}
\description{
Get info for 1 study
}
\examples{
info("2015-001314-10")
}
