% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cache}
\alias{read_cache}
\title{Read cache}
\usage{
read_cache(key, cache_file = ".eudract.rds")
}
\arguments{
\item{key}{a key to cache data}

\item{cache_file}{a file to store cached data}
}
\value{
Cached data if exist. Otherwise NULL
}
\description{
Read cache
}
