% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_studies}
\alias{search_studies}
\title{Search Studies in Eudract}
\usage{
search_studies(query, size = NULL)
}
\arguments{
\item{query}{a text query}

\item{size}{number of results to return}
}
\value{
a list of results
}
\description{
Search Studies in Eudract
}
\examples{
search_studies("dupilumab")
}
