% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_id}
\alias{validate_id}
\title{Validate eudract id}
\usage{
validate_id(eudract_id)
}
\arguments{
\item{eudract_id}{an eudract id to test}
}
\value{
TRUE or FALSE
}
\description{
Validate eudract id
}
