% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_cache}
\alias{write_cache}
\title{Write cache}
\usage{
write_cache(key, data, cache_file = ".eudract.rds")
}
\arguments{
\item{key}{a key to cache data}

\item{data}{a object to cache}

\item{cache_file}{a file to store cached data}
}
\value{
TRUE or FALSE
}
\description{
Write cache
}
